gopls: use new gomodcache index
This CL changes the default for 'importsSource' to 'gopls',
so that imports and unimpoted completions will use the GOMODCACHE
index.

One test had to be changed, and there is a benchmark, with some data.
(On unimported completions from the module cache, the new code is
more than 10 times faster, and always succeeds where the old code
sometimes failed.)

Change-Id: Ie5c8001ac1292498b72e5b42b51b4fcb06ab6fa9
Reviewed-on: https://go-review.googlesource.com/c/tools/+/678475
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>