gopls/internal/golang: add web-based "split package" tool
This CL contains a prototype of an interactive tool to help
you split a package into two or more independent
components whose dependency graph is acyclic.

The "source.splitPackage" code action opens a web browser
on a page that displays all the declarations in the current
package, grouped by file. You can create two or more
components, then assign declarations to components by
clicking checkboxes. As the assignment (or code) changes,
the tool displays the component dependency graph, along
with the symbol references that create that graph.
If there is a cycle in the dependency graph, it is flagged.

It doesn't actually move the declarations yet, but IMHO
that's actually a lesser problem than the analysis.

The component names and their assignments for each package
are saved persistently in the filecache.

+ integration test (using a Go client in lieu of JS+DOM),
  docs, relnote

Change-Id: I45d5e5d47950bcab988a51d314c5c41831562de3
Reviewed-on: https://go-review.googlesource.com/c/tools/+/679815
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Madeline Kalil <mkalil@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
Commit-Queue: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>