gopls/internal/golang: Hover: fix crash on alias to built-in Named
parseFull had an unstated precondition that pos was valid, and
this was violated for built-ins. The primary Hover logic rejects
built-ins earlier, but the logic for aliases to built-in Named
types did not. The fix is to make parseFull take an Object,
not a Pos, and support built-in objects.

Fixes golang/go#74351

Change-Id: I84cb3ac9f02236dff7bcaa7c4f78c761e6fc7aa5
Reviewed-on: https://go-review.googlesource.com/c/tools/+/683536
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>