gopls/internal/cmd: fix directory cleanup issue in flaky test
The gopls command uses a temporary directory to run. We need
to kill the gopls process before the test finishes so that
the testing framework can clean up the temp directory. The
call to goplsCmd.Kill() is asynchronous, so this might be
causing a race condition where the testing framework attempts
to cleanup the directory before the gopls process has been killed.
Adding a call to cmd.Wait() should fix this.

For golang/go#74257

Change-Id: I5334fced0eec402dbccf25164d7ddfda950737e6
Reviewed-on: https://go-review.googlesource.com/c/tools/+/683695
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>