internal/mcp: disable schema validation
Schema validation is causing breakage for some usage inide of Google.
While this may be a downstream bug, it is blocking the import of x/tools
into Google. Disable it for now, to unblock the import.

No tests failed when schema validation is removed. We need to add more
tests before re-enabling this feature.

Change-Id: I72905a93bfb0a2d81bae467d903f027e71719487
Reviewed-on: https://go-review.googlesource.com/c/tools/+/684935
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>