gopls/internal/golang: allow rename in doc comments
This CL enables the renaming of an identifier from within its
doc comment. It modifies the PrepareRename command to check
for renames inside doc comments and then changes the rename
target to be the identifier instead of the doc comment
so that all references to the identifier are updated.

Fixes golang/go#42301

Change-Id: I5110d805dbe499c3e4010a5bd313df2a51b9f296
Reviewed-on: https://go-review.googlesource.com/c/tools/+/686015
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Madeline Kalil <mkalil@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>