gopls/internal/golang: fix extract on anonymous structs
Extracting a code block that contains an anonymous struct
currently fails with a "parent nil" error. This CL allows
extracting of anonymous structs and fields. It also fixes
a bug where a locally defined type is used in but not
included in the extracted block by throwing an error
when the extraction is attempted.

Fixes golang/go#61496
Change-Id: Iabecb2efa136a1806917deff11a4d320118579f5
Reviewed-on: https://go-review.googlesource.com/c/tools/+/689196
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>