gopls/internal/filewatcher: fix race condition on watcher shutdown
The watcher' Close method was subject to a race condition that could
cause a panic on a closed internal error channel.

The Watcher.Close method waits for all watch-registration goroutines
to exit before closing the errs channel. However, due to a race it was
possible for another watch-registration task to be added concurrently
to the Wait, resulting in a send to a closed channel.

This change fixes the race by moving the watchers' Add call into
the addWatchHandle method, which is guarded by a mutex. This ensures
that the check for a closing watcher and the increment of the waitgroup
are atomic. Once the Close method acquires the mutex, no new watch
goroutines can be started, guaranteeing that the watchers' Wait method
will correctly account for all active goroutines and all the send
operations to internal error channel.

Fixes golang/go#74805

Change-Id: I741a4a3b02358ba45a8c51a1097ec67b59a71be4
Reviewed-on: https://go-review.googlesource.com/c/tools/+/691935
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Hongxiang Jiang <hxjiang@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>