gopls/internal/folding_range: allow more folding ranges
When the 'lineFoldingOnly' is set, the client ignores character
positions in folding ranges. (Vscode does this, eglot does not support
floating ranges at all.) For multiline ranges delimited by (), [], or {},
the existing code will not generate a folding range if there is anything
but whitespace after the opening (, [, or { on the same line.

This seems too restrictive, as it excludes ranges starting with
func foo() { // non-godoc comment
or
var x = []string{"a",
so this CL removes that test.

Fixes: golang/go#73735
Change-Id: Ie7aa5e735542b20a72e582e407b5d6ae2b8b2d29
Reviewed-on: https://go-review.googlesource.com/c/tools/+/694295
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>