[release-branch.0.25] internal/tokeninternal: move unsafe gopls code into gopls
In x/tools@v0.25.0, the tokeninternal package accessed the token.FileSet
unsafely, and used a trick to force the build to break if ever the size
of token.FileSet changed. This workaround was only necessary for gopls,
but was unfortunately implemented in a package that was reachable
through the public API of x/tools.

In go1.25, the size of the token.FileSet changed, breaking tools that
indirectly depended on this poisoned package. Fix this by moving the
gopls logic into gopls: the gopls build will still be broken, but that
doesn't matter as we only support the most recent version of gopls,
which is fixed.

For golang/go#74462

Change-Id: Ied3f7a615cab66d88ec6f4cfe8738f8429865993
Reviewed-on: https://go-review.googlesource.com/c/tools/+/697336
Reviewed-by: Peter Weinberger <pjw@google.com>
TryBot-Bypass: Robert Findley <rfindley@google.com>
Reviewed-on: https://go-review.googlesource.com/c/tools/+/697416
Auto-Submit: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>