[release-branch.0.24] gopls: fix the build with go1.25
Turn the compile-time error in AddExistingFiles into a runtime panic,
and avoid it by delegating to the FileSet.AddExistingFiles method on
go1.25.

Also disable broken tests. Most of the breakage is related to
go/packages or the importer, as there was an export data change that is
not ported to this branch.

for golang/go#74462

Change-Id: I430209b329ab88da676253e2bf5f66d1792078bd
Reviewed-on: https://go-review.googlesource.com/c/tools/+/697337
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Peter Weinberger <pjw@google.com>