gopls/completion: unimported completions with explicit package names
If a user types foo., and foo is not a package already imported in the
current file, the unimported completion code searches for the
package foo in a number of places. If it fails to find it, the new
code looks in the current module for an import like
import foo "package-path-for-bar"
and tries to satisfy the completion from package bar. If it can,
it will also add a suitable import statement.

Related: golang/go#40278 golang/go#74918
Change-Id: I5345a7d1a23fa240c468b9b21b2e9bcddf27c770
Reviewed-on: https://go-review.googlesource.com/c/tools/+/698578
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>