gopls/internal/analysis/modernize: use better qualifier for rangeint
The init and/or limit part of a for loop may use a non-basic type
like "type ID int" defined in some other package.
When converting such cases to a modernized range statement, the
types.RelativeTo Qualifier copies the full package path which results
in strange formatting. Use typesinternal.FileQualifier instead.

If the init part is created via a function call
where the function definition is in a different package than
the current one, then we may need to add a new import when
replacing the loop with a range statement (since the non-basic
type won't already be imported). However, it is
not safe to perform a modernization in this case, because
there may be other side effects from the function so removing
it to create the "for range" would result in a behavior change.

If the limit part is created via a function call that returns a
non-basic type, the init must use the non-basic type so the
import would already exist.

Fixes golang/go#74687

Change-Id: Ief21b9aefff9b6b961e8aa6e7e286783283f9d93
Reviewed-on: https://go-review.googlesource.com/c/tools/+/700398
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
Auto-Submit: Madeline Kalil <mkalil@google.com>