godoc: isolate module
We use the replace trick here so that the module can build from a
checkout but it can't build with the requirement. We'll have to update
the requirement once there's a version of tools with this module
carved out.

For golang/go#59056

Change-Id: I6a6a69649eafa7c2eb6e165ab922bc3972d577d0
Reviewed-on: https://go-review.googlesource.com/c/tools/+/700596
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Matloob <matloob@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>