cmd/godoc: isolate module
We use the replace trick here so that the module can build from a
checkout but it can't build with the requirement. We'll have to update
the requirement once there's a version of tools with the godoc and
cmd/godoc modules removed.

For golang/go#59056

Change-Id: I6a6a69647784a9f9df902593bc1675472408d0f0
Reviewed-on: https://go-review.googlesource.com/c/tools/+/700598
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Matloob <matloob@google.com>