gopls/internal/golang: OutgoingCalls: fix crash on error.Error
As always, we have a nil deref panic due to the lack of a package.
This CL fixes the crash, and permits OutgoingCalls to show
calls to built-ins with a package (such as unsafe.Slice),
but not those without a package (such as len or error.Error).
This is somewhat arbitrary.

Fixes golang/go#75230

Change-Id: I38a86a8a6eee080eb0e7e4f3907e07b5ffff9cdb
Reviewed-on: https://go-review.googlesource.com/c/tools/+/700600
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>