gopls/internal/analysis/modernize/rangeint: fix defer bug
The for i := range f() {...} transformation alters the final value
of i, so we cannot apply it when i is used after the loop.
However, we neglected a subtle edge case: when i is used
(syntactically) before the loop, by a defer statement.
In this case the actual access of i occurs dynamically
after the loop.

This CL rejects the transformation if i is used within an
enclosing defer statement that itself is within the scope of i.

Also, reject if i is a global variable.

+ tests

Change-Id: Iddcff138e2c1882433dee13396d51ba4cfc6f961
Reviewed-on: https://go-review.googlesource.com/c/tools/+/701715
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>