gopls/internal/util/cursorutil: add FirstEnclosing
A common pattern while using cursors is finding the first cursor that
points to a type of node from an iterator and then extracting the node
from the cursor.  In many cases, the cursor is never used again.

This change adds a helper function to grab a value and do the necessary
type conversion, while also returning the cursor if the user wants it.
Because there was not an obvious place to put it, a new utility package
has been added.

Change-Id: I2c5eac7d4bde7cd99fcd948018b73095a4b8020e
Reviewed-on: https://go-review.googlesource.com/c/tools/+/702775
Reviewed-by: Alan Donovan <adonovan@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>