go/analysis/unitchecker: redirect stdout to specified file
This change causes unitchecker to redirect its stdout to the
file specified by the build system, so that it can be separated
from stderr and printed cleanly (and atomically) in the output
of "go vet". The preserves the integrity of (e.g.) JSON even in
the case of partial writes, concurrent cmd/vet tasks, or
other junk appearing on stderr.

See CL 702815 for the corresponding change in the go command.
That CL includes an integration test, but must wait until
x/tools has been revendored in GOROOT.

For golang/go#75432

Change-Id: Ib1a13b5338b044e3357830d1f67b507505c819e9
Reviewed-on: https://go-review.googlesource.com/c/tools/+/702835
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Matloob <matloob@google.com>
Reviewed-by: Michael Matloob <matloob@golang.org>