gopls/internal/doc/generate: convert Analyzer.Doc to Markdown
The format of Analyzer.Doc fields is not specified, but in
practice all of gopls' analyzers (from x/tools and staticcheck)
use go/doc/comment form. Although this is similar to a subset
of Markdown, strictly it should be passed through the
comment.{Parser,Printer} machinery, for example to quote
Markdown metacharacters (*, [, etc) when they do not have
special meanings in the go/doc/comment subset.
This change adds this step to the doc generator.

A consequence is that ##, when used as a subheading in a
Go doc comment, will be escaped so that it appears literally.
This makes the rendering of analyzers.md consistent with
pkg.go.dev and gopls' internal doc viewer.
Unlike Markdown, Go doc comments have no concept of
subheadings (at least today), yet there are occasions
when a subheading is warranted, such as the docs for
internal/gofix. The meaning of a literal ## is clear enough,
even though it won't render as a heading nor appear in the ToC.

(Staticcheck's SA9004 analyzer also uses subheadings in its
doc string literal, but of the pre-go1.19 legacy variety.
However, internal/gofix cannot use this approach because,
in a doc comment, gofmt will helpfully convert it to a #
heading. Consequently, staticcheck gets to add subheadings
in the ToC, but our own analyzers do not.)

The output was verified in go doc, go doc -http (pkgsite),
and analyzers.md as rendered by go.dev/gopls/analyzers.

We should probably specify the form of Analyzer.Doc comments
more precisely.

Fixes golang/go#75199

Change-Id: I44f3e0fad9cdd5bdc77b34fa23147f4e0dce8a09
Reviewed-on: https://go-review.googlesource.com/c/tools/+/704735
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>