go/analysis/passes/printf: check anonymous functions too
This change causes the printf analyzer to consider anonymous
functions that are directly assigned to a variable, such as

   printf := func(format string, args ...any) { ... }

to be analyzed the same way as an ordinary function. This
allows mistakes calling local printf-wrapper functions to
be reported.

We denote printf-like functions by a types.Object (which
may be a *Func or, now, a *Var) throughout the algorithm.

This generalizes to local and global vars and struct fields,
even ones imported from another package, though since one
cannot export a fact for foreign symbols, the imported var
will only have the wrapper property within the unit that
encloses the assignment.

Also:
- a number of clarifications to identifier names, comments, etc:
- fix a couple of latent printf bugs in x/tools.

+ tests

Change-Id: I82c23dae4a682b574c2b0f4e6daa9be260ae47be
Reviewed-on: https://go-review.googlesource.com/c/tools/+/706635
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>