gopls/internal: prompt with full package path in preparerename
As part of implementing package move refactoring (rename on a package
that results in moving the package to a different directory), we
need to allow the user to provide us with a full package path. First,
we change PrepareRename to prompt the user with the full package path
instead of just the package name. For example, PrepareRename on package
golang would display "golang.org/x/tools/gopls/internal/golang" instead
of "golang".

For now, throw an error if the user attempts to rename a package to
a different directory.

Also update the marker test framework to handle directory renames.

For golang/go#57171

Change-Id: If2f7d1cfcbb9ce142f619b5ccb04843bca5d40f5
Reviewed-on: https://go-review.googlesource.com/c/tools/+/708655
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>