gopls/internal/golang: change hover handling from position to range
All the computation from hover handling changes from a position to
a range. If the hover request does not have a range, it is interpreted
as a range where start and end are the same position.

An additional validation is added to some utility function to verify
the syntax node of interest covers the range (both the start and end)
instead of single point previously.

Hover marker, by default, will send range information derived from
the source location.

Fixes an issue: gopls return rune info inside of a string literal even
the cursor is not inside of the rune.

For golang/go#69058

Change-Id: If1aaae454a4a6c8607bdc8758ccea2cd8157ccf4
Reviewed-on: https://go-review.googlesource.com/c/tools/+/709495
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>