go/analysis/passes/modernize: reflecttypefor: remove orphaned var
This CL fixes a bug in reflecttypefor that caused it to leave
a declaration for a local variable that was no longer used,
which is not legal.  Now, it deletes the variable:

- var zero T
- reflect.TypeOf(zero)
+ reflect.TypeFor[T]()

The bookkeeping is done by the new refactor.DeleteUnusedVars function.

Fixes golang/go#75767

Change-Id: If51af1500a63c968098b4a518b941315e826fcbf
Reviewed-on: https://go-review.googlesource.com/c/tools/+/710656
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: shuang cui <imcusg@gmail.com>
Auto-Submit: Alan Donovan <adonovan@google.com>