gopls/internalanalysis: DeleteStmt deletes comments
DeleteStmt is called with just the AST so it has trouble deciding if
a comment is associated with the statement to be deleted. For instance,
for /*A*/ init()/*B*/;/*C/cond()/*D/;/*E*/post() /*F*/ { /*G*/}
comment B and C are indistinguishable, as are D and E. That is, as the
AST does not say where the semicolons are, B and C could go either
with the init() or the cond(), so cannot be removed safely. The same
is true for D, E, and the post(). (And there are other similar cases.)
But the other comments can be removed as they are unambiguously
associated with the statement being deleted.

The most common case where the new code differs is that it removes
whole lines like
  stmt // comment

Change-Id: I2d0652892241aac3a80f1b0a0cafbfeec2e6274f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/710775
Reviewed-by: Alan Donovan <adonovan@google.com>
TryBot-Bypass: Peter Weinberger <pjw@google.com>