gopls/mod: find references to required modules
With this CL, users will be able to click on a required module in a
go.mod file, and see the import statements that reference its
packages.

In addition to integrating with LSP clients, this code avoids
reporting imports of nested modules.

Testing: there are marker tests, but self-contained. A lot
of natural uses will be to find modules from the module cache.
I tested by hand getting references from a go.mod file in
the tools repository. For instance, when searching for references to
golang/x/tools in gopls/go.mod, for marker_test.go it finds
the 8 imports of packages in golang.org/x/tools,
but none of the many golang.go/x/tools/gopls imports.

Fixes golang/go#75810

Change-Id: I08022dd29173a6d4ea8c728c7eb8e54a462a5889
Reviewed-on: https://go-review.googlesource.com/c/tools/+/713160
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>