gopls/internal/golang: typeDefinition: allow any expression
Previously, typeDefinition required an identifier with a type.
Now it permits any expression.

Also
- fix the bug in the type-switch edge case in the TODO comment.
- fix bug in marker test of @typedef; now it follows the shape
  of ordinary @def, and supports multiple locations.
  And is documented.

(The real motive for this change was to remove a use of
referencedObject, which needs to be melted down for golang/go#75975.)

+ tests

For golang/go#69058
For golang/go#75975

Change-Id: If8d29d3b920a51f12624ed512c29292bdbb4fbb4
Reviewed-on: https://go-review.googlesource.com/c/tools/+/713600
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hongxiang Jiang <hxjiang@golang.org>