go/ast/inspector: FindByPos: fix bug in FuncDecl.Type
The FuncDecl.{Name,Type} fields are an edge case in many
algorithms because their ranges overlap. That breaks
the assumption of FindByPos that the latest enclosing
node in the traversal is the innermost one.
Add logic to deal with this edge case, and a test.

Fixes golang/go#75997

Change-Id: I4ed2021d304d576a0e8667903c1f40e591631e4c
Reviewed-on: https://go-review.googlesource.com/c/tools/+/713920
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>