internal/astutil: add Range type
This CL defines a type Range that is simply a [start, end]
pair of token.Pos, and uses it in a number of places.

It is often convenient to treat the pair as a single
entity, and to have them satisfy the Node and
analysis.Range interfaces, and to provide handy operations
such as Contains.

Three other types were eliminated:
- analysisinternal.Range
- golang.posRange
- packagestest.Range

Change-Id: I00ca582c8e8ed77cc62e270e4fb98b2ae1def333
Reviewed-on: https://go-review.googlesource.com/c/tools/+/714401
Reviewed-by: Hongxiang Jiang <hxjiang@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>