gopls/internal/golang: Hover: show field info for embedded fields too
This CL removes the logic that caused Hover over an embedded field
to report on the type, not the field itself, which was unhelpful
when (e.g.) inspecting field size/align information.

The logic was originally common to Hover and Definition (where it
is still wanted), but was moved up into Hover itself in the
preceding CL of the stack of cleanups attached to this issue.

+ test

Fixes golang/go#75975

Change-Id: Ice74a430638f48349e300cf442685137a45a09ef
Reviewed-on: https://go-review.googlesource.com/c/tools/+/714701
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>