go/analysis/passes/modernize: newexpr: add //go:fix inline directives
This CL causes the newexpr suggested fix to add "inline" directives
on the function equivalent to new(expr) so that calls to it are
eventually melted away. This is safe because the inliner now checks
that the caller's Go version is at least as new as the callee's.

Updates golang/go#45624
Updates golang/go#75726

Change-Id: I36d27c7fa92965340da739ab71ed0cddff0d03ce
Reviewed-on: https://go-review.googlesource.com/c/tools/+/716561
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>