go/analysis/passes/modernize: fix illegal byte argument error produced in stringscut modernizer
This CL fixes a bug in the stringscut modernizer. Previously, an instance of strings.IndexByte(s, 'a') would become strings.Cut(s, 'a') with 'a' being an illegal argument because strings.Cut expects a string, not a byte.

If we examine an instance of strings.IndexByte and are about to suggest a modernization, we append an additional edit that transforms the byte literal or byte variable substr arg to a string literal.

For example, the following calls to strings.IndexByte would yield the substr arg:
strings.IndexByte(_, 'a') -> strings.Cut(_, "a")

strings.IndexByte(_, 55) -> strings.Cut(_, "7")

b := byte('a')
strings.IndexByte(_, b) -> strings.Cut(_, string(b))

For bytes.IndexByte, the second argument must be a converted to a []byte:

bytes.IndexByte(_, 'a') -> bytes.Cut(_, []byte{'a'})

Change-Id: I1a7670e95ebdffca5ce7a2a09a103f1a6fee6db3
Reviewed-on: https://go-review.googlesource.com/c/tools/+/716600
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>