go/analysis/internal/analysisflags: -fix: remove unused imports
This change defines analysisflags.FormatSourceRemoveImports, a
replacement for format.Source that removes imports that are no
longer needed in a modified file.

It requires the types.Package for the unmodified file so that it
can tell what name is declared by an existing non-renaming import.
Any name declared by an import that is not free in the rest of
the file is considered redundant, and is removed.

Individual analyzers typically remove an import when they can
tell that this is necessary, but when several fixes are composed,
and each removes the second-last reference to an import, the
problem must be solved holistically; thus the right place is
in the anlayzer driver.

The new function is used in unitchecker, single/multichecker,
and analysistest.

It is tested indirectly via tests of individual analyzers such
as inline and modernize.

Fixes golang/go#72035

Change-Id: I490e160c11efd44ac63f5c33dac7c1b94c425258
Reviewed-on: https://go-review.googlesource.com/c/tools/+/717802
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>