go/analysis/passes/modernize: stditerators: add reflect iters
This CL adds to the stditerators table 6 new iterators
in the reflect package. It also adds support for iter.Seq2,
since the loop must be "for _, x", not "for x".

(See CL 707356 for the addition of the actual iterators.)

+ tests

For golang/go#66631
For golang/go#75693

Change-Id: I59b4f55a45316ecf6a9dda180a7d1f0f9134d130
Reviewed-on: https://go-review.googlesource.com/c/tools/+/717620
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>