gopls/internal/mcp: add go_rename_symbol MCP tool
Adds a new MCP tool for renaming a symbol. Given a URI, a symbol name,
and a new name, it computes the edits necessary to rename that symbol
across the Go workspace. It borrows the logic in the
go_symbol_references tool to find the location of the symbol in
the file and then calls rename to fetch the necessary edits.

The idea to add an MCP tool for renaming comes from cl/694036, which
differs from this CL because it uses position information instead of
a symbol name as input.

Fixes golang/go#74887

Change-Id: I4a94d4d8208c6a0b106a5d63c23fcecb1c257cea
Reviewed-on: https://go-review.googlesource.com/c/tools/+/718120
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>