internal/diff/lcs: introduce line diffs
This CL modifies the lcs code to also do line diffs,
that is, diffs on string slices. Line diffs are presently
used in tests, but will be useful in the future to show the
effect of fixing and modernizers.

This is part of a plan to replace the diff/myers package
which does not produce minimal line edits,
but there are also places converting lcs diffs into line diffs
where the new code will be much faster.

Change-Id: Ia0f48d5fec2ebb7b314b71e3d58f5bab26ef8995
Reviewed-on: https://go-review.googlesource.com/c/tools/+/719160
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>