go/analysis/passes/modernize: forvar: handle "if v := v; cond {"
Some analyzers' (especially stditerator's) fixes result
in redundant v := v assignments, often as the "init" part of
an if statement. This change causes the forvar analyzer to
clean them up when the if statement is the sole statement
of the loop body.

+ test

For golang/go#76241
For golang/go#76240

Change-Id: I327d8726b0d2439a5cc55765acefdfa463282728
Reviewed-on: https://go-review.googlesource.com/c/tools/+/719360
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>