internal/refactor/inline: avoid expensive Info.Uses scan
This CL further optimizes (beyond the attempt in CL 717540)
the computation of whether a given PkgName is used only once.
The previous attempt reduced the scan of Info.Uses from
once per import to once per inline operation, but if
a //go:fix inline-annotated function is called a lot,
this is still a lot of work.

The Caller struct now allows a client to provide an optimized
implementation of the CountUses query. The inline analyzer
uses the pre-computed typeindex.Index to make it essentially
free.

The gopls inline code action does not provide it,
so it falls back to the slow implementation, but that
should be ok since it's only a single call site and the
operation is explicitly invoked. The "inline all" operation
deals with multiple callsites, but again it is explicitly
invoked. The key issue here is the cost of always-on
background analysis, not user-initiated code actions.

(I noticed from the profile that AST printing is also a
significant cost. I wonder if we should make this analyzer
use gopls' lazy-edits mechanism, at least when it is
running within gopls, to defer these costs. Or, better,
stop formatting and reparsing so much when computing
edits.)

Fixes golang/go#75773

Change-Id: I9a6f3275c524cc3841d5f1b3f8016509bacf7efe
Reviewed-on: https://go-review.googlesource.com/c/tools/+/720440
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>