go/analysis/passes/inline: add lazy-fix mode for gopls
Computation of inliner fixes can be expensive due to its
(unnecessary) strategy of formatting and parsing the file
repeatedly. Rather than fix that right now, this CL makes
the computation of fixes lazy when the analyzer is running
within gopls. An undocumented analyzer flag, set by gopls'
init, causes the analyzer to return an empty list of edits,
which triggers gopls to compute the edits lazily based on
the Diagnostic.Category string. As luck would have it, we
already have a lazy fix in place for the refactor.inline.call
Code Action, so we simply use that name again for the Category.

The change broke an MCP test that assumed that the inline
analyzer's fixes carry edits. There is a TODO in the
MCP code not to rely on this assumption, but rather than
do it, I rewrote the test to use the printf analyzer
instead.

For golang/go#75773

Change-Id: I5388e472e841cb23351fb9e90d8d7721dae3df07
Reviewed-on: https://go-review.googlesource.com/c/tools/+/721180
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>